Implemented middleware functionality

+++ routes.php
@@ -4,22 +4,22 @@ declare(strict_types=1);
 // Get routes
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingsController@index');
-$router->get('/listings/create', 'ListingsController@create');
-$router->get('/listings/edit/{id}', 'ListingsController@edit');
+$router->get('/listings/create', 'ListingsController@create', ['auth']);
+$router->get('/listings/edit/{id}', 'ListingsController@edit', ['auth']);
 $router->get('/listings/{id}', 'ListingsController@show');
 
 // Post routes
-$router->post('/listings', 'ListingsController@store');
+$router->post('/listings', 'ListingsController@store', ['auth']);
 
 // Delete routes
-$router->delete('/listings/{id}', 'ListingsController@destroy');
+$router->delete('/listings/{id}', 'ListingsController@destroy', ['auth']);
 
 // Put routes
-$router->put('/listings', 'ListingsController@update');
+$router->put('/listings', 'ListingsController@update', ['auth']);
 
 // Authentication routes
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
-$router->post('/auth/register', 'Usercontroller@store');
-$router->post('/auth/logout', 'UserController@logout');
-$router->post('/auth/login', 'UserController@authenticate');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
+$router->post('/auth/register', 'Usercontroller@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
